# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Lorem Ipsum",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/hr/applications": {
            "get": {
                "tags": [
                    "HR - Applications"
                ],
                "summary": "Mendapatkan daftar lamaran",
                "description": "List semua lamaran masuk dengan filter status atau vacancy.",
                "operationId": "11837f16da1227a567d40fff9bdfd5e0",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter berdasarkan status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "vacancy_id",
                        "in": "query",
                        "description": "Filter berdasarkan ID lowongan",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Pencarian berdasarkan nama atau email kandidat",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar lamaran berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Application"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/hr/applications/{id}": {
            "get": {
                "tags": [
                    "HR - Applications"
                ],
                "summary": "Melihat detail lamaran",
                "description": "Detail lamaran lengkap beserta data kandidat dan riwayat interview.",
                "operationId": "7a2e357d8a766ebf4a0f28be056d7400",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID lamaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data lamaran berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Lamaran tidak ditemukan"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/hr/applications/{id}/screening": {
            "patch": {
                "tags": [
                    "HR - Applications"
                ],
                "summary": "Pindah status ke screening",
                "description": "Pindahkan lamaran ke status screening — HR mulai mereview. Hanya untuk status applied.",
                "operationId": "99b81f622fd7221aa6dcbf80b0f63320",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID lamaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status lamaran berhasil diupdate ke screening",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lamaran dipindahkan ke screening."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Lamaran tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi status gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/hr/applications/{id}/reject": {
            "patch": {
                "tags": [
                    "HR - Applications"
                ],
                "summary": "Tolak lamaran",
                "description": "Tolak lamaran di tahap manapun (applied / screening / interview_scheduled).",
                "operationId": "2d1395fa17a6714638eae52bb3e12261",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID lamaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status lamaran berhasil diupdate ke rejected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lamaran berhasil ditolak."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Lamaran tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi status gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Application": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "candidate_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "vacancy_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "application_code": {
                        "type": "string",
                        "example": "APP-2024-ABCDEF"
                    },
                    "status": {
                        "type": "string",
                        "example": "applied",
                        "enum": [
                            "applied",
                            "screening",
                            "interview_scheduled",
                            "interview_done",
                            "hired",
                            "rejected"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Terjadi kesalahan."
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}